# perf: use parse instead of getMany for gjson